tests: removed unused deps and funcs
