+++ tests/Cargo.toml
@@ -6,18 +6,9 @@ license.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-anyhow = "1"
-kube = { version = "0.87", features = ["runtime", "derive", "ws"] }
-k8s-openapi = { version = "0.20", features = ["v1_25"] }
-curl = { version = "0.4", features = ["static-curl"] }
-rand = "0.8"
-tokio = { version = "1", features = ["rt", "macros", "process"] }
-
-
 [dev-dependencies]
-http = "1.1"
-tower = "0.5"
-hyper = "1.4"
+anyhow = "1"
+curl = { version = "0.4", features = ["static-curl"] }
 redis = { version = "0.25", features = ["tokio-comp"] }
 rumqttc = "0.24.0"
+tokio = { version = "1", features = ["rt", "macros", "process"] }
